requirements.txt
python-telegram-bot[webhooks]==20.7
aiohttp==3.9.1
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
pytz==2023.3
lxml==4.9.3
flask==3.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1 